scripts/requirements-photo-ai.txt
openai>=1.0.0
supabase
python-dotenv
